Add global exception handling and order validation
Introduces a global exception handler and standardized error response for API errors. Adds validation annotations to order request DTOs, refactors order creation to include user context and product validation, and updates quantity fields from Long to Integer for consistency. Also improves error handling in ProductService and updates the database schema accordingly.

+++ api/src/main/java/com/cc/api/order/dto/response/OrderDetailResponse.java
@@ -19,8 +19,8 @@ public class OrderDetailResponse {
 
     private ProductResponse product;
 
-    private Long quantity;
+    private Integer quantity;
 
     private BigDecimal subtotal;
-    
+
 }
